Manuellt lyft över specifications från 0.5-RB.

+++ specifications/src/main/java/se/inera/webcert/spec/TaBortFraga.groovy
@@ -1,22 +1,19 @@
 package se.inera.webcert.spec
 
 import se.inera.webcert.spec.util.RestClientFixture
-/**
- * Created by pehr on 9/23/13.
- */
-public class TaBortFraga extends RestClientFixture {
+
+class TaBortFraga extends RestClientFixture {
 
     String internReferens
     String externReferens
-
-    public void execute() {
+    String frageText
+    String svarsText
+    
+    def execute() {
         def restClient = createRestClient(baseUrl)
-
-        if (internReferens)
-            restClient.delete(path: "questions/${internReferens}")
-
-        if (externReferens)
-            restClient.delete(path: "questions/extern/${externReferens}")
+        if (internReferens) restClient.delete(path: "questions/${internReferens}")
+        if (externReferens) restClient.delete(path: "questions/extern/${externReferens}")
+        if (frageText) restClient.delete(path: "questions/frageText/${frageText}")
+        if (svarsText) restClient.delete(path: "questions/svarsText/${svarsText}")
     }
 }
-
